+++ GestionnaireAgendas/src/java/service/EvenementDAO.java
@@ -12,17 +12,17 @@ public interface EvenementDAO {
 
     public void delete (Evenement e);
 
-    public HashMap<Integer,Evenement> findAll ();
+    public HashMap<Long,Evenement> findAll ();
 
-    public Evenement findByPrimaryKey (int eventID);
+    public Evenement findByPrimaryKey (long eventID);
 
-    public HashMap<Integer,Evenement> findAfter (Date d);
+    public HashMap<Long,Evenement> findAfter (Date d);
 
-    public HashMap<Integer,Evenement> findBefore (Date d);
+    public HashMap<Long,Evenement> findBefore (Date d);
 
-    public HashMap<Integer,Evenement> findBetween (Date d1, Date d2);
+    public HashMap<Long,Evenement> findBetween (Date d1, Date d2);
 
-    public HashMap<Integer,Evenement> findByAgenda (int agendaID);
+    public HashMap<Long,Evenement> findByAgenda (long agendaID);
 
 }
